simpified BukAddPage
