+++ src/pages/BulkAddPage.tsx
@@ -1,1097 +1,40 @@
-import { useEffect, useRef, useState } from "react";
-import type { Dispatch, KeyboardEvent, SetStateAction } from "react";
-import { useNavigate, Link } from "react-router";
-import { format, parseISO, subDays } from "date-fns";
-import { ArrowLeft, Calendar, ChevronDown, Clock, Trash2 } from "lucide-react";
-import { toast } from "sonner";
+import { ArrowLeft } from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Switch } from "@/components/ui/switch";
-import { Select, SelectContent, SelectTrigger } from "@/components/ui/select";
-import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from "@/components/ui/alert-dialog";
-import {
-  CategorySelectValue,
-  CategoryOptionItems,
-} from "@/components/categories/CategorySelectOptions";
-import { TagChipList } from "@/components/expenses/TagChipList";
-import { useCategories } from "@/hooks/useExpenseData";
-import { useCurrency } from "@/contexts/CurrencyContext";
-import { getCurrentTime24 } from "@/lib/time";
-import { addExpensesBulk, getDescriptionSuggestions } from "@/db/expenseTrackerDb";
-import { getBulkDraft, saveBulkDraft, clearBulkDraft } from "@/db/bulkDraft";
-import type { BulkDraft, BulkDraftBlock, BulkDraftRow } from "@/db/bulkDraft";
-import type { Category, Expense } from "@/types/expense";
-import { cn } from "@/lib/utils";
-
-const SCALE_HINT_THRESHOLD = 100;
-const DRAFT_DEBOUNCE_MS = 500;
-
-type RowFieldErrors = { amount?: string; category?: string; tags?: string };
-type ExpenseDraft = Omit<Expense, "id" | "createdAt" | "updatedAt">;
-
-// ---------------------------------------------------------------------------
-// Pure helpers — row/block shape, validation, and the blocks[] state reducers.
-// Kept outside the component and split narrowly so each stays cheap to reason
-// about (and cheap for static analysis) on its own.
-// ---------------------------------------------------------------------------
-
-function isRowBlank(row: BulkDraftRow): boolean {
-  return row.value.trim() === "" && row.description.trim() === "" && row.tags.length === 0;
-}
-
-function blockHasContent(block: BulkDraftBlock): boolean {
-  return block.rows.some((r) => !isRowBlank(r));
-}
-
-function makeRow(category: string): BulkDraftRow {
-  return {
-    id: crypto.randomUUID(),
-    value: "",
-    category,
-    description: "",
-    tags: [],
-    time: null,
-    isAdhoc: false,
-  };
-}
-
-function makeBlock(date: string, category: string): BulkDraftBlock {
-  return { id: crypto.randomUUID(), date, collapsed: false, rows: [makeRow(category)] };
-}
-
-function blockTotals(block: BulkDraftBlock): { count: number; total: number } {
-  const nonBlank = block.rows.filter((r) => !isRowBlank(r));
-  return {
-    count: nonBlank.length,
-    total: nonBlank.reduce((sum, r) => sum + (Number(r.value) || 0), 0),
-  };
-}
-
-function pageTotals(blocks: BulkDraftBlock[]): { count: number; total: number } {
-  return blocks.reduce(
-    (acc, b) => {
-      const t = blockTotals(b);
-      return { count: acc.count + t.count, total: acc.total + t.total };
-    },
-    { count: 0, total: 0 },
-  );
-}
-
-function resolveTime(block: BulkDraftBlock, row: BulkDraftRow, today: string): string {
-  return row.time ?? (block.date === today ? getCurrentTime24() : "12:00");
-}
-
-function amountFormatError(trimmedValue: string): string | undefined {
-  if (trimmedValue === "") return "Amount is required";
-  if (Number.isNaN(Number(trimmedValue))) return "Amount is required";
-  return undefined;
-}
-
-function amountRangeError(amount: number): string | undefined {
-  if (amount <= 0) return "Must be positive";
-  if (amount > 10_000_000) return "Maximum 10,000,000";
-  return undefined;
-}
-
-function amountError(trimmedValue: string): string | undefined {
-  return amountFormatError(trimmedValue) ?? amountRangeError(Number(trimmedValue));
-}
-
-function validateRow(row: BulkDraftRow): RowFieldErrors {
-  return {
-    amount: amountError(row.value.trim()),
-    category: row.category === "" ? "Category required" : undefined,
-    tags: row.tags.length > 4 ? "Maximum 4 tags" : undefined,
-  };
-}
-
-function rowHasErrors(errors: RowFieldErrors): boolean {
-  return !!(errors.amount || errors.category || errors.tags);
-}
-
-function canAddTag(tags: string[], value: string): boolean {
-  return value !== "" && tags.length < 4 && !tags.includes(value);
-}
-
-function buildExpenseDraft(block: BulkDraftBlock, row: BulkDraftRow, today: string): ExpenseDraft {
-  return {
-    value: Number(row.value),
-    category: row.category,
-    description: row.description.trim() || undefined,
-    tags: row.tags,
-    date: block.date,
-    time: resolveTime(block, row, today),
-    isAdhoc: row.isAdhoc,
-  };
-}
-
-function flattenRows(blocks: BulkDraftBlock[]): { block: BulkDraftBlock; row: BulkDraftRow }[] {
-  return blocks.flatMap((block) => block.rows.map((row) => ({ block, row })));
-}
-
-interface InvalidRow {
-  blockId: string;
-  errors: RowFieldErrors;
-}
-
-interface PartitionResult {
-  toWrite: ExpenseDraft[];
-  invalidRows: Map<string, InvalidRow>;
-}
-
-/** Splits every row into a write-ready draft or a validation failure. Blank rows are
- *  silently dropped (neither list). */
-function partitionRows(blocks: BulkDraftBlock[], today: string): PartitionResult {
-  const toWrite: ExpenseDraft[] = [];
-  const invalidRows = new Map<string, InvalidRow>();
-
-  for (const { block, row } of flattenRows(blocks)) {
-    if (isRowBlank(row)) continue;
-    const errors = validateRow(row);
-    if (rowHasErrors(errors)) invalidRows.set(row.id, { blockId: block.id, errors });
-    else toWrite.push(buildExpenseDraft(block, row, today));
-  }
-
-  return { toWrite, invalidRows };
-}
-
-function firstInvalidEntry(
-  invalidRows: Map<string, InvalidRow>,
-): { rowId: string; blockId: string } | null {
-  const first = invalidRows.entries().next();
-  return first.done ? null : { rowId: first.value[0], blockId: first.value[1].blockId };
-}
-
-function toErrorMap(invalidRows: Map<string, InvalidRow>): Map<string, RowFieldErrors> {
-  return new Map(Array.from(invalidRows, ([id, v]) => [id, v.errors]));
-}
-
-function draftHasContent(draft: BulkDraft): boolean {
-  return draft.blocks.some(blockHasContent);
-}
-
-// blocks[] reducers — each takes the prior state and returns the next, so every handler in
-// the component is a one-line `setBlocks(bs => ...)` call.
-
-function updateRow(
-  rows: BulkDraftRow[],
-  rowId: string,
-  patch: Partial<BulkDraftRow>,
-): BulkDraftRow[] {
-  return rows.map((r) => (r.id !== rowId ? r : { ...r, ...patch }));
-}
-
-function updateBlockRow(
-  blocks: BulkDraftBlock[],
-  blockId: string,
-  rowId: string,
-  patch: Partial<BulkDraftRow>,
-): BulkDraftBlock[] {
-  return blocks.map((b) =>
-    b.id !== blockId ? b : { ...b, rows: updateRow(b.rows, rowId, patch) },
-  );
-}
-
-function addRowToBlock(
-  blocks: BulkDraftBlock[],
-  blockId: string,
-  row: BulkDraftRow,
-): BulkDraftBlock[] {
-  return blocks.map((b) => (b.id === blockId ? { ...b, rows: [...b.rows, row] } : b));
-}
-
-function removeRowFromBlock(
-  blocks: BulkDraftBlock[],
-  blockId: string,
-  rowId: string,
-  fallbackCategory: string,
-): BulkDraftBlock[] {
-  return blocks.map((b) => {
-    if (b.id !== blockId) return b;
-    const rows = b.rows.filter((r) => r.id !== rowId);
-    return { ...b, rows: rows.length > 0 ? rows : [makeRow(fallbackCategory)] };
-  });
-}
-
-function removeBlock(blocks: BulkDraftBlock[], blockId: string): BulkDraftBlock[] {
-  return blocks.filter((b) => b.id !== blockId);
-}
-
-function toggleBlockCollapse(blocks: BulkDraftBlock[], blockId: string): BulkDraftBlock[] {
-  return blocks.map((b) => (b.id === blockId ? { ...b, collapsed: !b.collapsed } : b));
-}
-
-function expandBlock(blocks: BulkDraftBlock[], blockId: string): BulkDraftBlock[] {
-  return blocks.map((b) => (b.id === blockId ? { ...b, collapsed: false } : b));
-}
-
-function setBlockDate(blocks: BulkDraftBlock[], blockId: string, date: string): BulkDraftBlock[] {
-  return blocks.map((b) => (b.id === blockId ? { ...b, date } : b));
-}
-
-function fillPlaceholderCategory(blocks: BulkDraftBlock[], categoryId: string): BulkDraftBlock[] {
-  return blocks.map((b) => ({ ...b, rows: fillRowCategory(b.rows, categoryId) }));
-}
-
-function fillRowCategory(rows: BulkDraftRow[], categoryId: string): BulkDraftRow[] {
-  return rows.map((r) => (r.category === "" ? { ...r, category: categoryId } : r));
-}
-
-function commitOnEnter(e: KeyboardEvent<HTMLInputElement>, onCommit: () => void) {
-  if (e.key !== "Enter") return;
-  e.preventDefault();
-  onCommit();
-}
-
-// ---------------------------------------------------------------------------
-// Custom hooks — each owns one cross-cutting concern's effects, isolated from
-// the page component's own body.
-// ---------------------------------------------------------------------------
-
-function pickDefaultCategoryId(categories: Category[]): string {
-  const others = categories.find((c) => c.name === "Others");
-  if (others) return others.id;
-  if (categories.length > 0) return categories[0].id;
-  return "";
-}
-
-function applyDefaultCategory(
-  categoryId: string,
-  setBlocks: Dispatch<SetStateAction<BulkDraftBlock[]>>,
-): void {
-  if (!categoryId) return;
-  setBlocks((bs) => fillPlaceholderCategory(bs, categoryId));
-}
-
-function useDefaultCategoryId(
-  categories: Category[],
-  setBlocks: Dispatch<SetStateAction<BulkDraftBlock[]>>,
-): string {
-  const defaultCategoryId = pickDefaultCategoryId(categories);
-  useEffect(
-    () => applyDefaultCategory(defaultCategoryId, setBlocks),
-    [defaultCategoryId, setBlocks],
-  );
-  return defaultCategoryId;
-}
-
-interface DraftPersistence {
-  pendingDraft: BulkDraft | null;
-  resumeDraft: () => void;
-  discardDraft: () => void;
-}
-
-function useBulkDraftPersistence(
-  blocks: BulkDraftBlock[],
-  setBlocks: Dispatch<SetStateAction<BulkDraftBlock[]>>,
-): DraftPersistence {
-  const [pendingDraft, setPendingDraft] = useState<BulkDraft | null>(null);
-  const [draftDecided, setDraftDecided] = useState(false);
-
-  useEffect(() => {
-    let cancelled = false;
-    getBulkDraft().then((draft) => {
-      if (cancelled) return;
-      if (draft && draftHasContent(draft)) setPendingDraft(draft);
-      setDraftDecided(true);
-    });
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-
-  useEffect(() => {
-    if (!draftDecided || pendingDraft) return;
-    const timer = setTimeout(() => void saveBulkDraft({ blocks }), DRAFT_DEBOUNCE_MS);
-    return () => clearTimeout(timer);
-  }, [blocks, draftDecided, pendingDraft]);
-
-  const resumeDraft = () => {
-    if (!pendingDraft) return;
-    setBlocks(pendingDraft.blocks);
-    setPendingDraft(null);
-  };
-
-  const discardDraft = () => {
-    setPendingDraft(null);
-    void clearBulkDraft();
-  };
-
-  return { pendingDraft, resumeDraft, discardDraft };
-}
-
-// ---------------------------------------------------------------------------
-// Presentational leaves — each renders one concern with no more than a single
-// guard clause or ternary of its own.
-// ---------------------------------------------------------------------------
-
-function RowFieldError({ message }: { message: string | undefined }) {
-  if (!message) return null;
-  return <p className="text-xs text-destructive">{message}</p>;
-}
-
-interface RowAmountCategoryProps {
-  row: BulkDraftRow;
-  categories: Category[];
-  currencySymbol: string;
-  errors: RowFieldErrors | undefined;
-  onUpdate: (patch: Partial<BulkDraftRow>) => void;
-  onRemove: () => void;
-  onCommit: () => void;
-  registerAmountRef: (el: HTMLInputElement | null) => void;
-}
-
-function RowAmountCategory({
-  row,
-  categories,
-  currencySymbol,
-  errors,
-  onUpdate,
-  onRemove,
-  onCommit,
-  registerAmountRef,
-}: RowAmountCategoryProps) {
-  const selectedCategory = categories.find((c) => c.id === row.category);
-
-  return (
-    <>
-      <div className="flex gap-2 items-start">
-        <div className="relative w-1/2">
-          <span className="absolute left-3 top-1/2 -translate-y-1/2 text-muted-foreground text-sm">
-            {currencySymbol}
-          </span>
-          <Input
-            ref={registerAmountRef}
-            aria-label="Amount"
-            inputMode="decimal"
-            placeholder="0"
-            className="pl-7"
-            value={row.value}
-            onChange={(e) => onUpdate({ value: e.target.value })}
-            onKeyDown={(e) => commitOnEnter(e, onCommit)}
-          />
-        </div>
-        <Select value={row.category} onValueChange={(v) => onUpdate({ category: v })}>
-          <SelectTrigger aria-label="Category" className="w-1/2">
-            <CategorySelectValue category={selectedCategory} placeholder="Category" />
-          </SelectTrigger>
-          <SelectContent>
-            <CategoryOptionItems categories={categories} />
-          </SelectContent>
-        </Select>
-        <button
-          type="button"
-          aria-label="Remove row"
-          onClick={onRemove}
-          className="p-2 text-muted-foreground hover:text-foreground"
-        >
-          <Trash2 className="h-4 w-4" />
-        </button>
-      </div>
-      <RowFieldError message={errors?.amount} />
-      <RowFieldError message={errors?.category} />
-    </>
-  );
-}
-
-function RowTagsToggle({
-  row,
-  onUpdate,
-}: {
-  row: BulkDraftRow;
-  onUpdate: (patch: Partial<BulkDraftRow>) => void;
-}) {
-  const [tagsOpen, setTagsOpen] = useState(row.tags.length > 0);
-  const [tagInput, setTagInput] = useState("");
-
-  const handleTagKey = (e: KeyboardEvent<HTMLInputElement>) => {
-    if (e.key !== "Enter") return;
-    e.preventDefault();
-    const value = tagInput.trim();
-    if (!canAddTag(row.tags, value)) return;
-    onUpdate({ tags: [...row.tags, value] });
-    setTagInput("");
-  };
-
-  if (!tagsOpen) {
-    return (
-      <button type="button" onClick={() => setTagsOpen(true)}>
-        + tag
-      </button>
-    );
-  }
-  return (
-    <Input
-      aria-label="Add tag"
-      placeholder="Add tag"
-      value={tagInput}
-      onChange={(e) => setTagInput(e.target.value)}
-      onKeyDown={handleTagKey}
-      className="h-7 w-24 text-xs"
-    />
-  );
-}
-
-interface RowTimeToggleProps {
-  block: BulkDraftBlock;
-  row: BulkDraftRow;
-  today: string;
-  onUpdate: (patch: Partial<BulkDraftRow>) => void;
-}
-
-function defaultRowTime(block: BulkDraftBlock, today: string): string {
-  return block.date === today ? getCurrentTime24() : "12:00";
-}
-
-function openTimePicker(input: HTMLInputElement | null): void {
-  if (!input) return;
-  input.focus();
-  try {
-    input.showPicker?.();
-  } catch {
-    // showPicker can throw outside a user gesture in some browsers — focus is enough of a
-    // fallback that the user can still open it manually.
-  }
-}
-
-function TimeToggleButton({ open, onClick }: { open: boolean; onClick: () => void }) {
-  return (
-    <button type="button" aria-label={open ? "Hide time" : "Set time"} onClick={onClick}>
-      <Clock className="h-3.5 w-3.5" />
-    </button>
-  );
-}
-
-function RowTimeInput({
-  open,
-  value,
-  onChange,
-  registerRef,
-}: {
-  open: boolean;
-  value: string;
-  onChange: (v: string) => void;
-  registerRef: (el: HTMLInputElement | null) => void;
-}) {
-  if (!open) return null;
-  return (
-    <Input
-      ref={registerRef}
-      type="time"
-      aria-label="Time"
-      className="h-7 w-24 text-xs"
-      value={value}
-      onChange={(e) => onChange(e.target.value)}
-    />
-  );
-}
-
-function RowTimeToggle({ block, row, today, onUpdate }: RowTimeToggleProps) {
-  const [timeOpen, setTimeOpen] = useState(row.time !== null);
-  const inputRef = useRef<HTMLInputElement | null>(null);
-
-  const handleToggle = () => {
-    const opening = !timeOpen;
-    setTimeOpen(opening);
-    if (opening) requestAnimationFrame(() => openTimePicker(inputRef.current));
-  };
-
-  return (
-    <>
-      <TimeToggleButton open={timeOpen} onClick={handleToggle} />
-      <RowTimeInput
-        open={timeOpen}
-        value={row.time ?? defaultRowTime(block, today)}
-        onChange={(v) => onUpdate({ time: v })}
-        registerRef={(el) => {
-          inputRef.current = el;
-        }}
-      />
-    </>
-  );
-}
-
-function RowAdhocToggle({
-  row,
-  onUpdate,
-}: {
-  row: BulkDraftRow;
-  onUpdate: (patch: Partial<BulkDraftRow>) => void;
-}) {
-  return (
-    <label className="flex items-center gap-1.5 ml-auto">
-      adhoc
-      <Switch
-        aria-label="adhoc"
-        checked={row.isAdhoc}
-        onCheckedChange={(v) => onUpdate({ isAdhoc: v })}
-      />
-    </label>
-  );
-}
-
-const DESCRIPTION_SUGGESTION_MIN_LENGTH = 2;
-const DESCRIPTION_SUGGESTION_LIMIT = 8;
-
-function filterDescriptions(all: string[], query: string): string[] {
-  const trimmed = query.trim().toLowerCase();
-  if (trimmed.length < DESCRIPTION_SUGGESTION_MIN_LENGTH) return [];
-  return all
-    .filter((d) => d.toLowerCase().includes(trimmed))
-    .slice(0, DESCRIPTION_SUGGESTION_LIMIT);
-}
-
-function RowDescriptionSuggestions({
-  show,
-  suggestions,
-  onSelect,
-}: {
-  show: boolean;
-  suggestions: string[];
-  onSelect: (value: string) => void;
-}) {
-  if (!show) return null;
-  return (
-    <div className="absolute z-50 w-full mt-1 bg-popover border border-border rounded-lg shadow-lg max-h-48 overflow-y-auto">
-      {suggestions.map((suggestion) => (
-        <button
-          key={suggestion}
-          type="button"
-          onClick={() => onSelect(suggestion)}
-          className="w-full text-left px-3 py-2 text-sm hover:bg-accent"
-        >
-          {suggestion}
-        </button>
-      ))}
-    </div>
-  );
-}
-
-interface RowDescriptionFieldProps {
-  value: string;
-  allDescriptions: string[];
-  onChange: (value: string) => void;
-  onKeyDown: (e: KeyboardEvent<HTMLInputElement>) => void;
-}
-
-function RowDescriptionField({
-  value,
-  allDescriptions,
-  onChange,
-  onKeyDown,
-}: RowDescriptionFieldProps) {
-  const [focused, setFocused] = useState(false);
-  const suggestions = focused ? filterDescriptions(allDescriptions, value) : [];
-
-  const selectSuggestion = (suggestion: string) => {
-    onChange(suggestion);
-    setFocused(false);
-  };
-
-  return (
-    <div className="relative">
-      <Input
-        aria-label="Description"
-        placeholder="Description"
-        value={value}
-        onChange={(e) => onChange(e.target.value)}
-        onKeyDown={onKeyDown}
-        onFocus={() => setFocused(true)}
-        onBlur={() => setTimeout(() => setFocused(false), 150)}
-      />
-      <RowDescriptionSuggestions
-        show={focused && suggestions.length > 0}
-        suggestions={suggestions}
-        onSelect={selectSuggestion}
-      />
-    </div>
-  );
-}
-
-interface RowCardProps {
-  block: BulkDraftBlock;
-  row: BulkDraftRow;
-  index: number;
-  today: string;
-  categories: Category[];
-  currencySymbol: string;
-  errors: RowFieldErrors | undefined;
-  allDescriptions: string[];
-  onUpdate: (patch: Partial<BulkDraftRow>) => void;
-  onRemove: () => void;
-  onCommit: () => void;
-  registerAmountRef: (el: HTMLInputElement | null) => void;
-  registerRowRef: (el: HTMLDivElement | null) => void;
-}
-
-function RowCard({
-  block,
-  row,
-  index,
-  today,
-  categories,
-  currencySymbol,
-  errors,
-  allDescriptions,
-  onUpdate,
-  onRemove,
-  onCommit,
-  registerAmountRef,
-  registerRowRef,
-}: RowCardProps) {
-  return (
-    <div
-      data-testid="bulk-row"
-      ref={registerRowRef}
-      className={cn("space-y-1.5 -mx-3 px-3 py-3", index % 2 === 1 && "bg-muted/40")}
-    >
-      <RowAmountCategory
-        row={row}
-        categories={categories}
-        currencySymbol={currencySymbol}
-        errors={errors}
-        onUpdate={onUpdate}
-        onRemove={onRemove}
-        onCommit={onCommit}
-        registerAmountRef={registerAmountRef}
-      />
-      <RowDescriptionField
-        value={row.description}
-        allDescriptions={allDescriptions}
-        onChange={(v) => onUpdate({ description: v })}
-        onKeyDown={(e) => commitOnEnter(e, onCommit)}
-      />
-      <div className="flex items-center gap-1 text-xs text-muted-foreground flex-wrap">
-        <RowTagsToggle row={row} onUpdate={onUpdate} />
-        <RowTimeToggle block={block} row={row} today={today} onUpdate={onUpdate} />
-        <RowAdhocToggle row={row} onUpdate={onUpdate} />
-      </div>
-      <TagChipList
-        tags={row.tags}
-        onRemoveTag={(t) => onUpdate({ tags: row.tags.filter((x) => x !== t) })}
-      />
-      <RowFieldError message={errors?.tags} />
-    </div>
-  );
-}
-
-function DateOption({ label, onClick }: { label: string; onClick: () => void }) {
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      className="w-full text-left px-2 py-1.5 text-sm rounded hover:bg-accent"
-    >
-      {label}
-    </button>
-  );
-}
-
-function blockDateLabel(block: BulkDraftBlock, today: string, yesterday: string): string {
-  if (block.date === today) return "Today";
-  if (block.date === yesterday) return "Yesterday";
-  return format(parseISO(block.date), "d MMM");
-}
-
-interface BlockDateControlProps {
-  block: BulkDraftBlock;
-  today: string;
-  yesterday: string;
-  onSetDate: (date: string) => void;
-}
-
-function BlockDateControl({ block, today, yesterday, onSetDate }: BlockDateControlProps) {
-  const [open, setOpen] = useState(false);
-
-  const choose = (date: string) => {
-    onSetDate(date);
-    setOpen(false);
-  };
-
-  return (
-    <Popover open={open} onOpenChange={setOpen}>
-      <PopoverTrigger asChild>
-        <button
-          type="button"
-          className="px-2.5 py-1 rounded-full text-xs font-medium bg-primary text-primary-foreground flex items-center gap-1 shrink-0"
-        >
-          <Calendar className="h-3 w-3" />
-          {blockDateLabel(block, today, yesterday)}
-        </button>
-      </PopoverTrigger>
-      <PopoverContent className="w-40 p-1" align="start">
-        <DateOption label="Today" onClick={() => choose(today)} />
-        <DateOption label="Yesterday" onClick={() => choose(yesterday)} />
-        <input
-          type="date"
-          aria-label="Custom date"
-          value={block.date}
-          onChange={(e) => choose(e.target.value)}
-          className="w-full h-8 text-xs border border-input rounded-md px-2 bg-background mt-1"
-        />
-      </PopoverContent>
-    </Popover>
-  );
-}
-
-function blockHeaderRadius(collapsed: boolean): string {
-  return collapsed ? "rounded-xl" : "rounded-t-xl";
-}
-
-interface BlockHeaderProps {
-  block: BulkDraftBlock;
-  today: string;
-  yesterday: string;
-  count: number;
-  total: number;
-  formatValue: (value: number) => string;
-  onToggleCollapse: () => void;
-  onSetDate: (date: string) => void;
-  onRemoveBlock: () => void;
-}
-
-function BlockHeader({
-  block,
-  today,
-  yesterday,
-  count,
-  total,
-  formatValue,
-  onToggleCollapse,
-  onSetDate,
-  onRemoveBlock,
-}: BlockHeaderProps) {
-  return (
-    <div
-      className={cn(
-        "flex items-center gap-2 p-3 bg-muted/30 flex-wrap",
-        blockHeaderRadius(block.collapsed),
-      )}
-    >
-      <button
-        type="button"
-        aria-label={block.collapsed ? "Expand day" : "Collapse day"}
-        onClick={onToggleCollapse}
-        className="shrink-0 text-muted-foreground hover:text-foreground"
-      >
-        <ChevronDown
-          className={cn("h-4 w-4 transition-transform", block.collapsed && "-rotate-90")}
-        />
-      </button>
-      <BlockDateControl block={block} today={today} yesterday={yesterday} onSetDate={onSetDate} />
-      <span className="ml-auto text-xs text-muted-foreground whitespace-nowrap">
-        {count} {count === 1 ? "entry" : "entries"} · {formatValue(total)}
-      </span>
-      <button
-        type="button"
-        aria-label="Remove day"
-        onClick={onRemoveBlock}
-        className="p-1 text-muted-foreground hover:text-foreground shrink-0"
-      >
-        <Trash2 className="h-4 w-4" />
-      </button>
-    </div>
-  );
-}
-
-interface BlockBodyProps {
-  block: BulkDraftBlock;
-  today: string;
-  categories: Category[];
-  currencySymbol: string;
-  errors: Map<string, RowFieldErrors>;
-  allDescriptions: string[];
-  onAddRow: () => void;
-  onUpdateRow: (rowId: string, patch: Partial<BulkDraftRow>) => void;
-  onRemoveRow: (rowId: string) => void;
-  registerAmountRef: (rowId: string, el: HTMLInputElement | null) => void;
-  registerRowRef: (rowId: string, el: HTMLDivElement | null) => void;
-}
-
-function BlockBody({
-  block,
-  today,
-  categories,
-  currencySymbol,
-  errors,
-  allDescriptions,
-  onAddRow,
-  onUpdateRow,
-  onRemoveRow,
-  registerAmountRef,
-  registerRowRef,
-}: BlockBodyProps) {
-  if (block.collapsed) return null;
-
-  return (
-    <div className="p-3 space-y-3">
-      {block.rows.map((row, index) => (
-        <RowCard
-          key={row.id}
-          block={block}
-          row={row}
-          index={index}
-          today={today}
-          categories={categories}
-          currencySymbol={currencySymbol}
-          errors={errors.get(row.id)}
-          allDescriptions={allDescriptions}
-          onUpdate={(patch) => onUpdateRow(row.id, patch)}
-          onRemove={() => onRemoveRow(row.id)}
-          onCommit={onAddRow}
-          registerAmountRef={(el) => registerAmountRef(row.id, el)}
-          registerRowRef={(el) => registerRowRef(row.id, el)}
-        />
-      ))}
-      <Button type="button" variant="ghost" size="sm" onClick={onAddRow}>
-        + Add transaction
-      </Button>
-    </div>
-  );
-}
-
-interface BlockCardProps extends BlockBodyProps {
-  yesterday: string;
-  formatValue: (value: number) => string;
-  onToggleCollapse: () => void;
-  onSetDate: (date: string) => void;
-  onRemoveBlock: () => void;
-}
-
-function BlockCard({
-  yesterday,
-  formatValue,
-  onToggleCollapse,
-  onSetDate,
-  onRemoveBlock,
-  ...bodyProps
-}: BlockCardProps) {
-  const { count, total } = blockTotals(bodyProps.block);
-
-  return (
-    <div data-testid="bulk-block" className="rounded-xl border border-border">
-      <BlockHeader
-        block={bodyProps.block}
-        today={bodyProps.today}
-        yesterday={yesterday}
-        count={count}
-        total={total}
-        formatValue={formatValue}
-        onToggleCollapse={onToggleCollapse}
-        onSetDate={onSetDate}
-        onRemoveBlock={onRemoveBlock}
-      />
-      <BlockBody {...bodyProps} />
-    </div>
-  );
-}
-
-function DraftResumeDialog({
-  pendingDraft,
-  onResume,
-  onDiscard,
-}: {
-  pendingDraft: BulkDraft | null;
-  onResume: () => void;
-  onDiscard: () => void;
-}) {
-  const resumeRef = useRef<HTMLButtonElement>(null);
-  if (!pendingDraft) return null;
-  const { count } = pageTotals(pendingDraft.blocks);
-
-  return (
-    <AlertDialog open>
-      <AlertDialogContent
-        data-testid="bulk-draft-prompt"
-        onEscapeKeyDown={(e) => e.preventDefault()}
-        onOpenAutoFocus={(e) => {
-          e.preventDefault();
-          resumeRef.current?.focus();
-        }}
-      >
-        <AlertDialogHeader>
-          <AlertDialogTitle>
-            Resume your unsaved draft ({count} {count === 1 ? "entry" : "entries"})?
-          </AlertDialogTitle>
-          <AlertDialogDescription>
-            Entries from your last visit are still on this device. Discarding removes them
-            permanently.
-          </AlertDialogDescription>
-        </AlertDialogHeader>
-        <AlertDialogFooter>
-          <AlertDialogCancel onClick={onDiscard}>Discard</AlertDialogCancel>
-          <AlertDialogAction ref={resumeRef} onClick={onResume}>
-            Resume
-          </AlertDialogAction>
-        </AlertDialogFooter>
-      </AlertDialogContent>
-    </AlertDialog>
-  );
-}
-
-function ScaleHint({ totalRowCount }: { totalRowCount: number }) {
-  if (totalRowCount <= SCALE_HINT_THRESHOLD) return null;
-  return (
-    <p className="text-xs text-muted-foreground">
-      That's a lot of rows —{" "}
-      <Link to="/settings/data/import-csv" className="underline">
-        CSV import
-      </Link>{" "}
-      is faster for big backfills.
-    </p>
-  );
-}
-
-function BulkFooter({
-  entryCount,
-  grandTotal,
-  formatValue,
-  onSave,
-  disabled,
-}: {
-  entryCount: number;
-  grandTotal: number;
-  formatValue: (value: number) => string;
-  onSave: () => void;
-  disabled: boolean;
-}) {
-  return (
-    <div className="fixed bottom-16 inset-x-0 lg:bottom-0 border-t border-border bg-background/95 backdrop-blur p-3 flex items-center justify-between gap-3 z-60">
-      <span className="text-sm text-muted-foreground">
-        {entryCount} {entryCount === 1 ? "entry" : "entries"} · {formatValue(grandTotal)}
-      </span>
-      <Button type="button" onClick={onSave} disabled={disabled}>
-        Save all
-      </Button>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Page
-// ---------------------------------------------------------------------------
+import { useBulkAddController } from "@/hooks/useBulkAddController";
+import { DraftResumeDialog } from "@/components/bulk-add/DraftResumeDialog";
+import { BlockCard } from "@/components/bulk-add/BlockCard";
+import { ScaleHint, BulkFooter } from "@/components/bulk-add/BulkFooter";
 
 export default function BulkAddPage() {
-  const navigate = useNavigate();
-  const categories = useCategories();
-  const { currency, formatValue } = useCurrency();
-
-  const today = format(new Date(), "yyyy-MM-dd");
-  const yesterday = format(subDays(new Date(), 1), "yyyy-MM-dd");
-
-  const [blocks, setBlocks] = useState<BulkDraftBlock[]>(() => [makeBlock(today, "")]);
-  const [rowErrors, setRowErrors] = useState<Map<string, RowFieldErrors>>(new Map());
-  const [saving, setSaving] = useState(false);
-  const lastCategoryRef = useRef("");
-  const amountRefs = useRef(new Map<string, HTMLInputElement>());
-  const rowRefs = useRef(new Map<string, HTMLDivElement>());
-  const [allDescriptions, setAllDescriptions] = useState<string[]>([]);
-
-  useEffect(() => {
-    // Suggestions are optional — a read failure must not surface as an unhandled rejection.
-    getDescriptionSuggestions(undefined, 500)
-      .then(setAllDescriptions)
-      .catch(() => setAllDescriptions([]));
-  }, []);
-
-  const defaultCategoryId = useDefaultCategoryId(categories, setBlocks);
-  const { pendingDraft, resumeDraft, discardDraft } = useBulkDraftPersistence(blocks, setBlocks);
-
-  const isDirty = blocks.some(blockHasContent);
-  const totalRowCount = blocks.reduce((n, b) => n + b.rows.length, 0);
-  const { count: entryCount, total: grandTotal } = pageTotals(blocks);
-
-  const clearRowError = (rowId: string) => {
-    setRowErrors((errs) => {
-      if (!errs.has(rowId)) return errs;
-      const next = new Map(errs);
-      next.delete(rowId);
-      return next;
-    });
-  };
-
-  const handleUpdateRow = (blockId: string, rowId: string, patch: Partial<BulkDraftRow>) => {
-    if (patch.category) lastCategoryRef.current = patch.category;
-    setBlocks((bs) => updateBlockRow(bs, blockId, rowId, patch));
-    clearRowError(rowId);
-  };
-
-  const handleAddRow = (blockId: string) => {
-    const newRow = makeRow(lastCategoryRef.current || defaultCategoryId);
-    setBlocks((bs) => addRowToBlock(bs, blockId, newRow));
-    requestAnimationFrame(() => amountRefs.current.get(newRow.id)?.focus());
-  };
-
-  const handleRemoveRow = (blockId: string, rowId: string) => {
-    setBlocks((bs) =>
-      removeRowFromBlock(bs, blockId, rowId, lastCategoryRef.current || defaultCategoryId),
-    );
-  };
-
-  const handleAddBlock = () => {
-    const newBlock = makeBlock(today, lastCategoryRef.current || defaultCategoryId);
-    setBlocks((bs) => [...bs, newBlock]);
-    requestAnimationFrame(() => amountRefs.current.get(newBlock.rows[0].id)?.focus());
-  };
-
-  const handleRemoveBlock = (blockId: string) => {
-    const block = blocks.find((b) => b.id === blockId);
-    if (!block) return;
-    if (blockHasContent(block) && !window.confirm("Remove this day and all its entries?")) return;
-    setBlocks((bs) => removeBlock(bs, blockId));
-  };
-
-  const handleToggleCollapse = (blockId: string) =>
-    setBlocks((bs) => toggleBlockCollapse(bs, blockId));
-  const handleSetDate = (blockId: string, date: string) =>
-    setBlocks((bs) => setBlockDate(bs, blockId, date));
-
-  const handleBack = () => {
-    if (isDirty && !window.confirm("Discard your unsaved entries?")) return;
-    navigate(-1);
-  };
-
-  const commitSave = async (toWrite: ExpenseDraft[]) => {
-    setSaving(true);
-    try {
-      const n = await addExpensesBulk(toWrite);
-      await clearBulkDraft();
-      toast.success(`${n} expense${n === 1 ? "" : "s"} added`);
-      navigate("/");
-    } catch {
-      toast.error("Couldn't save — nothing was written");
-      setSaving(false);
-    }
-  };
-
-  const handleSave = async () => {
-    if (saving) return;
-
-    const { toWrite, invalidRows } = partitionRows(blocks, today);
-    setRowErrors(toErrorMap(invalidRows));
-
-    const invalid = firstInvalidEntry(invalidRows);
-    if (invalid) {
-      setBlocks((bs) => expandBlock(bs, invalid.blockId));
-      requestAnimationFrame(() => {
-        rowRefs.current.get(invalid.rowId)?.scrollIntoView({ behavior: "smooth", block: "center" });
-      });
-      return;
-    }
-
-    if (toWrite.length === 0) return;
-    await commitSave(toWrite);
-  };
+  const {
+    categories,
+    currency,
+    formatValue,
+    today,
+    yesterday,
+    blocks,
+    rowErrors,
+    saving,
+    pendingDraft,
+    allDescriptions,
+    allTags,
+    totalRowCount,
+    entryCount,
+    grandTotal,
+    resumeDraft,
+    discardDraft,
+    handleUpdateRow,
+    handleAddRow,
+    handleRemoveRow,
+    handleAddBlock,
+    handleRemoveBlock,
+    handleToggleCollapse,
+    handleSetDate,
+    handleBack,
+    handleSave,
+    registerAmountRef,
+    registerRowRef,
+  } = useBulkAddController();
 
   return (
     <div className="px-4 py-6 pb-44 lg:pb-28 max-w-2xl mx-auto space-y-4">
@@ -1123,6 +66,7 @@ export default function BulkAddPage() {
           categories={categories}
           currencySymbol={currency.symbol}
           allDescriptions={allDescriptions}
+          allTags={allTags}
           formatValue={formatValue}
           errors={rowErrors}
           onSetDate={(date) => handleSetDate(block.id, date)}
@@ -1131,14 +75,8 @@ export default function BulkAddPage() {
           onAddRow={() => handleAddRow(block.id)}
           onUpdateRow={(rowId, patch) => handleUpdateRow(block.id, rowId, patch)}
           onRemoveRow={(rowId) => handleRemoveRow(block.id, rowId)}
-          registerAmountRef={(rowId, el) => {
-            if (el) amountRefs.current.set(rowId, el);
-            else amountRefs.current.delete(rowId);
-          }}
-          registerRowRef={(rowId, el) => {
-            if (el) rowRefs.current.set(rowId, el);
-            else rowRefs.current.delete(rowId);
-          }}
+          registerAmountRef={registerAmountRef}
+          registerRowRef={registerRowRef}
         />
       ))}
